Extend Toolbar instead of creating ElementToolbar
