+++ src/editor/SlateEditor.jsx
@@ -25,7 +25,7 @@ import isHotkey from 'is-hotkey';
 
 import './less/editor.less';
 
-import { ElementToolbar } from './ui/ElementToolbar';
+import Toolbar from './ui/Toolbar';
 
 const handleHotKeys = (editor, event, config) => {
   let wasHotkey = false;
@@ -247,11 +247,11 @@ class SlateEditor extends Component {
                 <InlineToolbar editor={editor} className={className} />
                 {Object.keys(slate.elementToolbarButtons).map((t) => {
                   return (
-                    <ElementToolbar elementType={t}>
+                    <Toolbar elementType={t}>
                       {slate.elementToolbarButtons[t].map((Btn) => {
                         return <Btn editor={editor} />;
                       })}
-                    </ElementToolbar>
+                    </Toolbar>
                   );
                 })}
               </>
